libs/radar - refactoring TempProfile class, updating apps to use new version of class

+++ codebase/apps/radar/src/Dsr2CP2Mom/Dsr2CP2Mom.cc
@@ -473,7 +473,7 @@ int Dsr2CP2Mom::_processField(const string &fieldName,
     bytesSent =
       _pProductSocket->writeDatagram((const char*)outBuf.getPtr(),
 				     outBuf.getLen());
-    if(bytesSent == outBuf.getLen()) {
+    if(bytesSent == (int) outBuf.getLen()) {
       return 0;
     }
     umsleep(100);
